sm1-de-assignment-1/feature-engineering/requirements.txt
pandas
google-cloud-storage
numpy
flask
requests
